t1479: add a section on OIDC
someone building a piece of OMS integrating with the OIDC server
will find everything they need and nothing they don't.

note that OIDC's original Setup Guide is here in source, but not
included in the build.

all of the links referenced ought to be refactored, using standard
notation and even the sphinx.ext.extlinks extension.