t1373: add a simple tutorial on remote editing files from windows
the idea is that you have a VM, either local to the developer's host
or out in the cloud, and changes are sync'd to the VM as you edit and
save them while editing them in Windows.

This should be cleaned up, formatted, and otherwise made proper before
merging.. then squash the changes together and merge!