t2319: Overhaul the documented release process
This is a massive overhaul, and aims to be as complete as possible, in
terms of coverage of the release process.

Prerequisites, various lists of responsibilities, steps to do in the
process, and examples of all configurations, updates, and commands to
run are included. Where necessary, both manual and automated processes
are documented.